Recursively resolve the repository path when provided a directory.
If the log format is not specified and 'logfile' is a directory,
recursively look for a version control repository.

This approach allows for something strange like someone who having an
svn repository inside a git repository (in which case it would pick
the svn directory as it would encounter that first).

Based off a prototype implementation by Jörg Bachmann.