author instantiation may crash if GetText() returns NULL
Gource will crash/dump core (tested on Windows, Linux and Mac OS X) if authorE->GetText() returns NULL.

This occured consistently with data generated from an SVN dump which was converted from a CVS repository, which in turn was originally converted from RCS files, some of which were 30 years old. Author information for check-ins was not always available, causing the commit user names to be invalid.

This is a quick fix at best.