Use built-in GLM version to fix compatibility issues (dependency removed).
The current GLM 0.9.9 pre-release alpha contains a change removing the
default initialization of vectors with no way of getting the old behavior.

This issue is compounded by Debian (and now Ubuntu) packaging this version of
GLM despite it being an alpha and not backwards compatible. I expect a lot of
other software will be affected by this decision as well.

For the time being I have decided to use a fixed version of GLM 0.9.8.5 including
a fix to compile with minor versions of GCC >= 7.3 which has not been included in
an official release yet.