Fixed a bug in the log file format detection affecting custom logs.
After the first commit in a log is read to check the file format the log
it is rewound to the beginning, however the last line read was still buffered
so it would return the buffered line next not the first line of the log. As
custom log entries are all on one line this would erroneously be parsed as the
first commit.