Decouple the order changes are applied from commit timestamps
Treat the order commits appear in the log file as canon rather
than make assumptions about the order of the timestamps being the
order the changes should be applied.

E.g. a file wont be removed if the previous action against it has
a higher commit_id than the removal action.